pyproject.toml
[tool.poetry]
name = "marketplace-managering-bot"
version = "0.1.0"
description = ""
authors = ["Artemio Panzini <panferovartemv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.6"
python-dotenv = "^1.0.0"
aiohttp = "^3.9.1"
aiohttp-retry = "^2.8.3"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
